Disable all buttons in the toolbar except Snapshot and Inventory
This required more changes to
indra/newview/skins/default/xui/en-us/panel_toolbar.xml
than adding 'moonworld="true"': that results in the two
buttons using up the whole window width.
So, I set the width of the buttons to a fixed size of 100 pixels now,
and have them follow the right side of the window.